\name{patccat_classify}
\alias{patccat_classify}
\title{Classify patent claims (alias for fn.patccat)}
\description{
Idiomatic alias for \code{\link{fn.patccat}}; identical arguments and return value.
}
\usage{
patccat_classify(data, ...)
}
\arguments{
\item{data}{A data frame of claim lines (see \code{\link{fn.patccat}}).}
\item{...}{Passed to \code{\link{fn.patccat}}.}
}
\value{See \code{\link{fn.patccat}}.}
\seealso{\code{\link{fn.patccat}}}
